backend/pyproject.toml
[project]
name = "spacefit-backend"
version = "0.1.0"
description = "SpaceFit AI Backend - Commercial Real Estate Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",

    # Browser automation
    "playwright>=1.41.0",

    # Data processing
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",
    "authlib>=1.3.0",

    # Rate limiting
    "slowapi>=0.1.9",

    # AI
    "anthropic>=0.40.0",

    # Payments
    "stripe>=8.0.0",

    # Date utilities
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
